requirement.txt
appdirs==1.4.4
asgiref==3.3.4
certifi==2020.12.5
distlib==0.3.1
Django==3.2.3
django-stdimage==5.3.0
django-taggit==1.4.0
filelock==3.0.12
Pillow==8.2.0
pipenv==2020.11.15
psycopg2==2.8.6
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
virtualenv==20.4.7
virtualenv-clone==0.5.4
